Adding ADT for Stations graph and some other changes
Stations graph is a graph which dedicates one node for ecery station. This node itself is a structure. The info is as follows:

1. Structure: StationNode
This is a node for every station. It has details about that station.
It also has 2 Vectors which maintain the other stations to which it is connected apart from the distance and DangerValue.

2. Structure: StructStations
This is basically an array of StationNodes. put into a structure just for more structure :D

Other changes:
1. Added a Tester.c just to check the input format for how the Station ADT was designed for.
2. Added "_vec" to all the names of functions in the vector ADT because names like resize, del, etc are common and can be confused with.

The end